requirements.txt
python-dotenv
pandas
matplotlib
seaborn
sqlalchemy
psycopg2-binary
rapidfuzz
argparse
